requirements.txt
blis==0.7.3
bootstrap4==0.1.0
catalogue==1.0.0
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
cymem==2.0.4
de-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-2.3.0/de_core_news_sm-2.3.0.tar.gz
dominate==2.6.0
gunicorn==20.0.4
Flask==1.1.2
Flask-Bootstrap4==4.0.2
Flask-Login==0.5.0
Flask-WTF==0.14.3
idna==2.10
iteration-utilities==0.11.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
MarkupSafe==1.1.1
murmurhash==1.0.4
nltk==3.5
numpy==1.19.4
oauthlib==3.1.0
plac==1.1.3
psycopg2-binary==2.8.6
preshed==3.0.4
pycparser==2.20
regex==2020.11.13
requests==2.25.0
six==1.15.0
spacy==2.3.4
srsly==1.0.4
thinc==7.4.3
tqdm==4.54.1
urllib3==1.26.2
visitor==0.1.3
wasabi==0.8.0
Werkzeug==1.0.1
WTForms==2.3.3
